Add InstagramPosts component and refactor Section1
